move messageq tests to go1.9
